wfc: performance: System checks.
Performance of well-formedness checks of large systems is improved.
1. In the test for recursiveness of a system, the calculation of all
subsystems is memoized by storing the data in an alist.
2. In the test for cyclic bindings in a system, the collection of all
reuired instances is calculate once in an alist, and re-used in the rest
of the test.

Co-authored-by: Jan Nieuwenhuizen <janneke@gnu.org>

* gaiag/wfc.scm (sub-systems-): New method.  Split worker out from...
(sub-systems): ... rename from subsystems.  Add memoization.
(all-required): Add 'required-alist' parameter to avoid re-calculation.
(cyclic-bindings): Calculate required-alist once and re-use it in
further processing.