parse: performance: Pre-calculate newlines for Location.
Location is stored in the ast as (line,col), so has to be transformed
from the position resulted from the peg parser.  This transformation was
taking 80% of parsing time.  Performance has been improved by
pre-calculating the positions of all newlines in the source.
Furthermore, calculations are only done for <locationed> nodes.

* gaiag/parse/ast.scm (parse-tree->ast): Calculate newlines and use
these in 'line-number' and 'column-number' functions, called only for
<locationed>.  Use 'tree-collect-filter' to optimize collecting imports.