simulate: Support compliance in blocking context.
The compliance was disabled for blocking contexts, as compliance can
only be checked after a system-run-to-completion, i.e., when the
component is unblocked again.  This patch keeps the trail while in
blocking context and postpones the compliance check until
system-run-to-completion.

* dzn/vm/goops.scm (<end-of-trail>): Inherit from <status> instead of
<error>.
* dzn/vm/report.scm (initial-error-message): Remove dead code.
(final-error-messages): Only display info for <end-of-trail> when AST is
set, i.e. when we stop midway a run-to-completion.
(report): Do not exit; return status.
* dzn/commands/simulate.scm (main): Instead, exit here.
* dzn/simulate.scm (check-provides-compliance): For blocking context,
use first event on trail.
(run-component): Remove.  Move compliance check...
(run-sut): ...here.  Change signature to take pc+blocked trace instead
of pc.
(run-trail): Update accordingly: For blocking context, do not report
intermediate trail; instead hand it back to run-sut.
* test/all/local.mk (XFAIL_TESTS): Remove compliance_blocking_out.