verify: Add deterministic check of interfaces.
* dzn/verify/pipeline.scm (in-out:lpsconstelm->lpsparelm): Rename to...
(in-out:lps->lpsparelm): ...this.
(in-out:lpsparelm->aut): Rename to...
(in-out:lps->aut): ...this.
(in-out:aut->aut-weak-trace): Change to procedure.  For interfaces, map
"model-name"'silent to tau.
(in-out:aut-dpweak-bisim->aut-failures): Rename to...
(in-out:aut->aut-failures): ...this.
(in-out:aut-dpweak-bisim->verify-interface): Rename to...
(in-out:aut->verify-interface): ..this.
(in-out:aut->verify-interface-nondet): New variable.
(in-out.pipeline): Update accordingly.  Use weak-trace instead of
dpweak-bisim for LTS of interfaces.
(report-fail): Add reporting of non-deterministic interfaces.
(mcrl2:verify-interface-asserts): Run both verify-interface and
verify-interface-nondet.  Add handling of results of determistic check
on interfaces.
* test/all/hello_nondet/META,
test/all/hello_silent/META: Disable verify.
* test/all/*/baseline/verify/*: Update verify baseline.