complete: Use uniform naming.
* dzn/parse/complete.scm (at-location?): Rename to...
(tree:at-location?): ...this.
(after-location?): Rename to...
(tree:after-location?): ...this.
(around-location?): Rename to...
(tree:around-location?): ...this.
(before-location?): Rename to...
(tree:before-location?): ...this.
(context:event-names): Rename to...
(complete:event-names): ...this.
(context:interface-names): Rename to...
(complete:interface-names): ...this.
(context:port-event-names): Rename to...
(complete:port-event-names): ...this
(context:trigger-names): Rename to...
(complete:trigger-names): ...this
(context:action-names): Rename to...
(complete:action-names): ...this.
(tree:function-names): Rename to...
(complete:function-names): ...this.
(context:enum-value-names): Rename to..
(complete:enum-literal-names): ...this.
(tree:int-value-names):  Rename to...
(complete:int-literal-names): ...this.
(tree:enum-field-test): Rename to...
(complete:field-test-names): ...this.
(context:type-value-names): Rename to...
(complete:type-literal-names): ...this.
(complete-enum-literal): Rename to...
(complete:enum-literal): ...this.
(complete-literal): Rename to...
(complete:literal): ...this.
(complete-for-type): Rename to...
(complete:for-type): ...this.
(context:complete-for-type): Rename to...
(complete:type): ...this.
(complete-function): Rename to...
(complete:function): ...this.
(complete-on): Rename to...
(complete:on): ...this.
(complete:boolean-expressions): Rename to...
(complete:boolean-expression): ...this.
(instance-ports): Move into...
(instance-ports-alist): ...here.
(complete-other-end-point): Rename to...
(complete:other-end-point): ...this.
(behaviour-top): Rename to...
(complete:behaviour): ...this.
(context:on-type): Rename to...
(.type-of-on): ...this.
(context:statements): Rename to...
(complete:statements): ...this.
(system-top): Rename to...
(complete:system): ...this.
(context:complete): Rename to...
(complete:root): ...this.
And update all callers.