verify: Add memoization of LTS for interface non-determinism check.
* dzn/verify/pipeline.scm (in-out:dzn->aut-dpweak-bisim-cached): New
procedure.
(in-out:aut->aut-weak-trace): Change "model-name"'silent to optional and inevitable.
(pretty-verify-pipeline): For pipeline debug commands, hide the dzn
command for the "verify-interface" pipelines.
(unmemoized-verify-pipeline): For pipeline debug commands, insert an
extra line-continuation and pipe symbol for the "aut-dpweak-bisim"
output.
(verify-pipeline-wrapper): New procedure.
(memoizing-verify-pipeline): Use it.
(verify-pipeline): Use it to avoid pipeline memoization when running the
pipeline in debug mode.
(mcrl2:verify-interface-asserts): Remove the check for the deterministic
result.