simulate: Fix split-arrows for blocking_system2,3.
Zip used to prepend the port trace prefix before the component trace.
For system traces with multiple blocking components, a port trace may
need to inserted mid-trace: just before the component handles the
trigger.

* dzn/ast.scm (ast:equal?)[trigger]: New specialization.
* dzn/simulate.scm (zip): Add trigger parameter.  Use it to find proper
location in the component trace to insert the port trace.  Also, do not
insert the port trace twice.  Do not synthesize an action twice.  Be
more strict in matching the component action with the port action; also
consider runtime instance.
(check-provides-compliance): Update caller.  Only reset PC for external
triggers in a blocked context.  When resetting event for a blocked PC
and a requires trigger, use the last trigger-return to find the matching
trigger.