vm: Avoid deadlock in state diagram through leakage of locals.
On a blocked thread, the popping of local variables was postponed to
just before the return after the release.  As the state diagram uses a
hash of the state, as an unintended consequence the state of a blocked
thread included local variables.  This caused for states to artificially
differ where they should have been the same, leading to deadlocks.

* dzn/vm/step.scm (step)[block]: In a blocking context, pop locals.
[end-of-on, trigger-return]: In a blocking context, skip pop locals.