dzn: Drop the global -j,--json option.
The --json option is a remnant of the pre-2.10 cloud-based
implementation.  It was only used by dzn parse and dzn traces.

On dzn traces, it printed the names of the traces to stdout after all
traces were generated; dzn --verbose traces will now do that as progress
on stderr.

The irregular "working-directory" in the AST and the move to guile-json4
have broken dzn --json parse in two ways.

* dzn/script.scm (parse-opts): Remove json option.
* dzn/commands/parse.scm (main): Remove attempt to produce json output.
* dzn/commands/traces.scm (model->traces): Remove json option.  Pass
verbose? to...
* dzn/lts.scm (lts->traces): ...here, using the new #:verbose?
parameter.  Likewise, pass it to..
(generate-trace): ..here, using the new #:verbose? parameter.  When
verbose?, print name of trace file to stderr.
* doc/dezyne.texi (Invoking dzn): Remove it.