test: collateral_blocking_bridges: Add proxy.
A component that blocks collaterally at the edge of a system is not
observable in the trace.  Only once it is released, the first event of
the trace is printed.

By adding a proxy between the edge of the system and the collaterally
blocking component, the collaterally blocking thread is observable when
it starts running.

* test/all/collateral_blocking_bridges/collateral_blocking_bridges.dzn
(top_proxy): New component.
(collateral_blocking_bridges): Add it below top.
* test/all/collateral_blocking_bridges/trace: Rename to...
* test/all/collateral_blocking_bridges/trace2: ...this.
* test/all/collateral_blocking_bridges/trace,
* test/all/collateral_blocking_bridges/trace1: New files.
* test/all/collateral_blocking_bridges/main.cc: Cater for all traces.
* test/all/collateral_blocking_bridges/main.cs: Likewise.