test: collateral_blocking_multiple_provides: Add proxies.
A component that blocks collaterally at the edge of a system is not
observable in the trace.  Only once it is released, the first event of
the trace is printed.

By adding a proxy between the edge of the system and the collaterally
blocking component, the collaterally blocking thread is observable when
it starts running.

* test/all/collateral_blocking_multiple_provides/collateral_blocking_multiple_provides.dzn
(proxy_left, proxy_right): New components.
(collateral_blocking_multiple_provides): Add them.