test: collateral_blocking_reorder: Add proxy.
A component that blocks collaterally at the edge of a system is not
observable in the trace.  Only once it is released, the first event of
the trace is printed.

By adding a proxy between the edge of the system and the collaterally
blocking component, the collaterally blocking thread is observable when
it starts running.

* test/all/collateral_blocking_reorder/collateral_blocking_reorder.dzn
(proxy_release): New component.
(Block): Rename to...
(blocked): ...this.
(collateral_blocking_reorder): Update accordingly and add proxy_release.
* test/all/collateral_blocking_reorder/trace,
test/all/multiple_parallel_blocking/trace: Update accordingly.