simulate: Allow blocking enabled events from free provides port.
This enables the interleaving of events from a free provides port when a
system is blocked on a blocking requires port, and collaterally blocking
on a component that is blocked in such a way.

* dzn/vm/runtime.scm (runtime:port)[runtime:component]: Change
specialization to...
[runtime:instance]: ...this.
* dzn/vm/util.scm (rtc-labels): Only include <rtc> labels when PC is
switchable.
(blocked-on-boundary-entry?): Cater for event #f.
(blocked-on-boundary-collateral-release, blocked-on-boundary-provides?:
New methods.
(collateral-block): Also consider being blocked-on-boundary?.
* dzn/vm/run.scm (mark-collateral-blocked): Cater for systems with
multiple provides ports.
(run-to-completion*): Only block if port is set.  Use
blocked-on-boundary-provides to determine illegal-trace.
(run-to-completion*-context-switch): Use
blocked-on-boundary-collateral-release to release collaterally blocked
coroutines.
* dzn/vm/step.scm (begin-step)[runtime-instance]: Also report
<blocked-error> if boundary port is already engaged.
* dzn/vm/step.scm (step-action-down): Also collaterally block if
port is blocked-on-boundary?.
* dzn/simulate.scm (end-report): Skip refusal check for blocked trace.
(run-trail): Also consider blocked-on-boundary? for blocked traces.