c++: Lookup operator << via ADL.
The C++ community appears to have settled on the policy that overloaded
operators should be placed in the same namespace as the type for which
they are overloaded, instead of the global namespace.  Examples of this
are the fact that 2.17.0 does not compile on most if not all of the
visual studio compilers.  Another example can be found in the google
test framework described here:
https://github.com/google/googletest/commit/0555b0eacbc56df1fd762c6aa87bb84be9e4ce7e.
And a third example is on stackoverflow:
https://stackoverflow.com/questions/3623631/where-should-non-member-operator-overloads-be-placed.

* dzn/templates/c++/enum-to-string@enum: Place operator << in overloaded
type namespace; place to_cstr and to_string in dzn namespace.
* dzn/templates/c++/main-event-map-valued-provides-in@trigger,
* dzn/templates/c++/main-port-connect-in@trigger: Prefix with dzn.
* dzn/templates/c++/string-to-enum@enum: Place to_"type" from string dzn
namespace.
* dzn/templates/c++/main@component: Move to ..
* runtime/c++/dzn/runtime.hh: Here and move with to_string into dzn namespace.
* test/all/blocking_shell/c++/main.cc: Update to_* with dzn:: accordingly.