test: Support META `model' flag.
The new `model' flag in META can be used to override the name of the
test's model, whereas the old `model' flag--now `component'--is used
for traces and verify system->component trickery in the absence of
system traces.

* test/dzn/dzn.scm (model?): New procedure.
(compenent?): Use it as fall-back.
(run-code, run-simulate-trace): Use it to solve long-standing FIXME.