verify: Improve performance of unreachable code check.
Improve performance of unreachable code check by modelling tags
as normale transitions and remove tags from the makreel source code
of interfaces when they are included for makreel code of component.

* dzn/ast/normalize.scm (tag-imperative-blocks): Simplify and unify
with e.g. add-defer-end.
* dzn/lts.scm (remove-tag-loops): Remove...
(assert-livelock): ...remove call to it.
(deadlock-nodes): ...remove call to it.
* dzn/code.scm (%no-tags-interface?): Add new parameter.
* dzn/code/scmackerel/makreel.scm <sm:tag>: New sm:node for tagging
statements.
(print-sm:tag): New function for displaying sm:tag node, don't
display tag if tag is part of an interface and in the context of a
component.
(ast->process <model>): Use sm:tag for tagging.
(port-processes): Remove blocking of tags.
* dzn/code/language/makreel.scm (root-> o): Set parameter
%no-tags-interface? to true when top model is a component.

* test/all/*: Change error traces to current reported error traces.