DRAFT verify: Deprecate -a,--all for new -k,--keep-going.
Also refactor implementation.

* dzn/commands/verify.scm (parse-opts): Add -k,--keep-going.  Mention
deprecation of -a,--all.
(main): Issue a deprecation warning when using -a,--all.  Pass
new #:keep-going parameter to...
* dzn/verify/pipeline.scm (verification:verify): ...here.  Add
new #:keep-going? parameter.  Refactor to use that instead of
command-line:get.  Pass it to...
(mcrl2:verify): ...here.  Likewise.  Pass it to...
(mcrl2:verify-component): ...here, and...
(mcrl2:verify-interface): ...here, and...
(mcrl2:verify-component-asserts): ...here, and...
(mcrl2:verify-interface-asserts): ...here, and...
(reduce-or): ...here.
* test/dzn/dzn.scm (run-verify): Use --keep-going.
* test/bin/update.sh: Likewise.
* doc/dezyne.texi (Invoking dzn verify): Document it.