vm: Make <on> observable instead of <initial-compound>.
* dzn/vm/util.scm (make-implicit-illegal): Rewrite to make an <on> with
a <declarative-illegal> statement.
* dzn/vm/step.scm (step <initial-compound>): Update accordingly.
(step <program-counter> <declarative-illegal>): Handle it in new
specialization.
* dzn/vm/compliance.scm (zip, check-provides-compliance,
check-provides-compliance+, update-shared-state): Use <on>
instead of <initial-compound>.
* dzn/vm/report.scm (pc-event? <initial-compound>): Change to...
(pc-event? <on>): ...this.
(pc->event <initial-compound>): Change to...
(pc-event <on>): ...this.
(pc-component? <initial-compound>): Change to...
(pc-component? <on>): ...this.
(pc->component <initial-compound>): Change to...
(pc-component <on>): ...this.
(pc-arrow? <initial-compound>): Change to...
(pc-arrow? <on>): ...this.
(pc->arrow <initial-compound>): Change to...
(pc-arrow <on>): ...this.
(complete-split-arrows-pcs): : Use <on> instead of <initial-compound>.