Revert "test: data_full: Add tests for <extern> data as return and reply."
Extending the overly complicated data_full test is highly undesirable;
the ROADMAP has listed the intend to split it for many years now.  Also,
it's not canonicalized.

Because C# does not support reply-data, adding that feature to data_full
would mean that the data path was no longer under test for C#.

This reverts commit b136b6ab0164cbf2d012348ecbf45be0134011f3.